Remove redundant FileTreeIterator.java Remove PathComparator.java Move Iterators to dedicated package
